Prefactoring of CRS related functions
